Extracting logic out of table to support the creation of the Dataset class
