+++ synapseclient/api/entity_factory.py
@@ -20,7 +20,7 @@ from synapseclient.core.exceptions import (
 
 if TYPE_CHECKING:
     from synapseclient import Synapse
-    from synapseclient.models import File, Folder, Project, Table
+    from synapseclient.models import Dataset, File, Folder, Project, Table
 
 
 async def get_from_entity_factory(
@@ -32,7 +32,7 @@ async def get_from_entity_factory(
     download_file: bool = True,
     download_location: str = None,
     follow_link: bool = False,
-    entity_to_update: Union["Project", "File", "Folder", "Table"] = None,
+    entity_to_update: Union["Project", "File", "Folder", "Table", "Dataset"] = None,
     *,
     synapse_client: Optional["Synapse"] = None,
 ) -> Union["Project", "File", "Folder"]:
@@ -353,7 +353,15 @@ async def _cast_into_class_type(
         if not entity_to_update:
             entity_to_update = Table()
         entity = entity_to_update.fill_from_dict(
-            synapse_table=entity_bundle["entity"], set_annotations=False
+            entity=entity_bundle["entity"], set_annotations=False
+        )
+    elif entity["concreteType"] == concrete_types.DATASET_ENTITY:
+        if not entity_to_update:
+            from synapseclient.models import Dataset
+
+            entity_to_update = Dataset()
+        entity = entity_to_update.fill_from_dict(
+            entity=entity_bundle["entity"], set_annotations=False
         )
     else:
         raise ValueError(
