+++ synapseclient/core/constants/concrete_types.py
@@ -66,6 +66,7 @@ FOLDER_ENTITY = "org.sagebionetworks.repo.model.Folder"
 LINK_ENTITY = "org.sagebionetworks.repo.model.Link"
 PROJECT_ENTITY = "org.sagebionetworks.repo.model.Project"
 TABLE_ENTITY = "org.sagebionetworks.repo.model.table.TableEntity"
+DATASET_ENTITY = "org.sagebionetworks.repo.model.table.Dataset"
 
 # upload requests
 MULTIPART_UPLOAD_REQUEST = "org.sagebionetworks.repo.model.file.MultipartUploadRequest"
